Stack values now of type double instead of int.

+++ lab3-reverse-polish-calc/stack.c
@@ -4,7 +4,7 @@
 #include "stack.h"
 
 typedef struct node {
-    int data;
+    double data;
     struct node *next;
 } *node_t;
 
@@ -23,17 +23,17 @@ void stack_free(stack_t stack) {
     free(stack);
 }
 
-void stack_push(stack_t stack, int value) {
+void stack_push(stack_t stack, double value) {
     node_t new_top = (node_t) malloc(sizeof(struct node));
     new_top -> data = value;
     new_top -> next = stack -> top;
     stack -> top = new_top;
 }
 
-int stack_pop(stack_t stack) {
+double stack_pop(stack_t stack) {
     assert(stack -> top != NULL);   /*! Precondition: stack must not be empty */
     node_t top_old = stack -> top;
-    int value = top_old -> data;
+    double value = top_old -> data;
     node_t top_new = top_old -> next;
     stack -> top = top_new;
     free(top_old);
